Fixed duplicate named test
