+++ Tests.Unit.Presentation.Web/DomainServices/QueryByRightsHolderItInterfaceTest.cs
@@ -2,12 +2,13 @@
 using System.Linq;
 using Core.DomainModel.ItSystem;
 using Core.DomainModel.Organization;
+using Core.DomainServices.Queries.Interface;
 using Tests.Toolkit.Patterns;
 using Xunit;
 
 namespace Tests.Unit.Presentation.Web.DomainServices
 {
-    public class QueryByRightsHolderTest : WithAutoFixture
+    public class QueryByRightsHolderItInterfaceTest : WithAutoFixture
     {
         [Fact]
         public void Apply_Returns_Items_With_RightsHolder_Uuid_Match()
@@ -57,7 +58,7 @@ namespace Tests.Unit.Presentation.Web.DomainServices
             };
 
             var input = new[] { excludedWrongUuid, matched, excludedNoExhibit, excludedNoRightsHolder }.AsQueryable();
-            var sut = new Core.DomainServices.Queries.Interface.QueryByRightsHolder(correctId);
+            var sut = new QueryByRightsHolder(correctId);
 
             //Act
             var result = sut.Apply(input);
@@ -67,27 +68,6 @@ namespace Tests.Unit.Presentation.Web.DomainServices
             Assert.Same(matched, itInterface);
         }
 
-
-        [Fact]
-        public void Apply_Returns_Items_With_Uuid_Match()
-        {
-            //Arrange
-            var correctId = A<Guid>();
-            var incorrectId = A<Guid>();
-            var matched = new ItSystem { BelongsTo = new Organization() { Uuid = correctId } };
-            var excludedNoRightsHolder = new ItSystem { BelongsTo = null };
-            var excludedWrongUuid = new ItSystem { BelongsTo = new Organization { Uuid = incorrectId } };
-
-            var input = new[] { excludedWrongUuid, matched, excludedNoRightsHolder }.AsQueryable();
-            var sut = new Core.DomainServices.Queries.ItSystem.QueryByRightsHolder(correctId);
-
-            //Act
-            var result = sut.Apply(input);
-
-            //Assert
-            var itSystem = Assert.Single(result);
-            Assert.Same(matched, itSystem);
-        }
     }
 }
