+++ Tests.Unit.Presentation.Web/Tests.Unit.Presentation.Web.csproj
@@ -208,9 +208,9 @@
     <Compile Include="DomainServices\QueryByBusinessTypeTest.cs" />
     <Compile Include="DomainServices\QueryByIdsTest.cs" />
     <Compile Include="DomainServices\QueryByNumberOfUsagesTest.cs" />
-    <Compile Include="DomainServices\QueryByRightsHolderTest.cs" />
+    <Compile Include="DomainServices\QueryByRightsHolderItSystemTest.cs" />
+    <Compile Include="DomainServices\QueryByRightsHolderItInterfaceTest.cs" />
     <Compile Include="DomainServices\QueryByTaskRefTest.cs" />
-    <Compile Include="DomainServices\QueryByRightsHolderTest.cs" />
     <Compile Include="DomainServices\QueryByRightsHolderIdsTest.cs" />
     <Compile Include="DomainServices\QueryByUuidTest.cs" />
     <Compile Include="DomainServices\QueryByIdTest.cs" />
